+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+
+* Hot fix for Bitfinex balances.
+
 # v2.3.2
 
 * Need wallet type for Bitfinex balance and adderss fetching.